requirements.txt
torch
torchvision
transformers
datasets
scikit-learn
matplotlib
tqdm
pillow
numpy
opencv-python
kaggle
opendatasets
onnx
onnxruntime
